+++ cmd/workload/get/cmd.go
@@ -75,9 +75,9 @@ Example:
 	outputformat.AddFlag(cmd, &outputFormat)
 	idargs.AddDirFlag(cmd)
 
-	telemetry.TrackWith(cmd, func(_ *cobra.Command, _ []string) map[string]any {
+	telemetry.TrackWith(cmd, func(_ *cobra.Command, args []string) map[string]any {
 		return map[string]any{
-			"workload_id":        ref.ID,
+			"workload_id":        idargs.TelemetryID(ref, args),
 			"workload_id_source": ref.Source,
 			"output_format":      string(outputFormat),
 		}